Include Expand fix for ONNX conformance tests failure:
https://github.com/microsoft/onnxruntime/pull/5854

ONNX conformance results:
`Summary: Total=3889, Passed=3604, Failed=0, Blocked=0, Not Run=0, Skipped=285`

+++ include/onnxruntime/core/providers/nnapi/nnapi_provider_factory.h
@@ -6,8 +6,8 @@
 
 // NNAPIFlags are bool options we want to set for NNAPI EP
 // This enum is defined as bit flats, and cannot have negative value
-// To generate a unsigned long nnapi_flags for using with OrtSessionOptionsAppendExecutionProvider_Nnapi below,
-//   unsigned long nnapi_flags = 0;
+// To generate an uint32_t nnapi_flags for using with OrtSessionOptionsAppendExecutionProvider_Nnapi below,
+//   uint32_t nnapi_flags = 0;
 //   nnapi_flags |= NNAPI_FLAG_USE_FP16;
 enum NNAPIFlags {
   NNAPI_FLAG_USE_NONE = 0x000,
@@ -28,7 +28,7 @@ enum NNAPIFlags {
 extern "C" {
 #endif
 
-ORT_API_STATUS(OrtSessionOptionsAppendExecutionProvider_Nnapi, _In_ OrtSessionOptions* options, unsigned long nnapi_flags);
+ORT_API_STATUS(OrtSessionOptionsAppendExecutionProvider_Nnapi, _In_ OrtSessionOptions* options, uint32_t nnapi_flags);
 
 #ifdef __cplusplus
 }
